Add exit section with link if closed Fixes #2745
